1- Add permission.

+++ lib/app/modules/notifications/views/widgets/notifications_buttons_widget.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:toby_bills/app/components/button_widget.dart';
 import 'package:toby_bills/app/components/icon_button_widget.dart';
+import 'package:toby_bills/app/core/utils/user_manager.dart';
 import 'package:toby_bills/app/core/values/app_colors.dart';
 import 'package:toby_bills/app/modules/notifications/controllers/notifications_controller.dart';
 
@@ -10,6 +11,7 @@ class NotificationsButtonsWidget extends GetView<NotificationsController> {
 
   @override
   Widget build(BuildContext context) {
+    final permission = UserManager().user.userScreens["customeraddnotice"]!;
     return Padding(
       padding: const EdgeInsets.symmetric(horizontal: 20).copyWith(top: 15.0),
       child: Row(
@@ -43,17 +45,25 @@ class NotificationsButtonsWidget extends GetView<NotificationsController> {
               return Row(
                 mainAxisSize: MainAxisSize.min,
                 children: [
-                  ButtonWidget(text: "حفظ", onPressed: () => controller.saveNotification()),
+                  if((permission.edit ?? false) || controller.invoice.value?.id == null)
+                    const SizedBox(width: 5),
+                  if((permission.edit ?? false) || controller.invoice.value?.id == null)
+                    ButtonWidget(text: "حفظ", onPressed: () => controller.saveNotification()),
                   const SizedBox(width: 5),
-                  ButtonWidget(text: "جديد", onPressed: () => controller.newInvoice()),
+                  if((permission.add ?? false))
+                    const SizedBox(width: 5),
+                  if((permission.add ?? false))
+                    ButtonWidget(text: "جديد", onPressed: () => controller.newInvoice()),
                   if(controller.notification.value?.id != null)
                     Row(
                       mainAxisSize: MainAxisSize.min,
                       children: [
                         const SizedBox(width: 5),
                         ButtonWidget(text: "طباعة قيد", onPressed: () => controller.printGeneralJournal(context)),
-                        const SizedBox(width: 5),
-                        ButtonWidget(text: "حذف", onPressed: () => controller.deleteNotification()),
+                        if((permission.delete ?? false) && controller.invoice.value?.id != null)
+                          const SizedBox(width: 5),
+                        if((permission.delete ?? false) && controller.invoice.value?.id != null)
+                          ButtonWidget(text: "حذف", onPressed: () => controller.deleteNotification()),
                       ],
                     ),
                   const SizedBox(width: 5),
